Batch device delete operations.
Summary:
This commit introduces batching for device delete operations. When the
EAS code invokes del_device() for a certain account, we have to make a
first round-trip to collect all the keys (account:folder) for that
account and, for each key, perform a device delete operation. By
batching these operations, we avoid multiple rount-trips.

Test Plan: Sync an EAS account.

Reviewers: khamidou, kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D935