Leverage the heartbeats infrastructure for Contacts and Events.
Summary:
This revision introduces heartbeats for Contacts and Events sync.
Since with "classic" sync, contacts and events aren't assigned a folder and
folder ID as in EAS sync, we use ad-hoc folder IDs (-1 and -2) for contacts
and events, respectively.

NOTE: this revision removes a personalized event poll frequency for Outlook.com
accounts since we were not even synchronizing events for those accounts. If we
discover we need something different than the default we will reintroduce it.

Test Plan: Make sure heartbeats flows from greenlets to the Redis datastore.

Reviewers: khamidou, emfree

Reviewed By: emfree

Subscribers: spang

Differential Revision: https://review.inboxapp.com/D893