Add a script to clear an account's heartbeat status.
Summary:
This revision adds a script to clear an account's heartbeat status. It takes an
account ID (as in our DB) as an input and may optionally take an hostname, a
port, and a database number in case you don't want to use the local
configuration.

Test Plan:
Sync an account and try to remove it's heartbeat status. Double-check that the
account's heartbeat status is removed by executing bin/inbox-syncstatus.

Reviewers: emfree, kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D971