README.md: add instructions to create own OAuth Client ID
Currently, the user is authenticating their email account to a shared OAuth Client ID, to which everyone has the client ID and the secret to. This seems insecure. This change updates the README.md to add instructions on how to create one's own OAuth Client ID and add it to secrets.yml.